+++ tests/testthat/test-ifelse.R
@@ -14,6 +14,13 @@ test_that("basic behavior of ifelse() works", {
 
   expect_equal(
     test |>
+      mutate(y = base::ifelse(x1 == 'a', "foo", "bar")) |>
+      pull(y),
+    c("foo", "foo", "bar", "foo", "bar")
+  )
+
+  expect_equal(
+    test |>
       mutate(y = ifelse(x1 == 'a', x3, x1)) |>
       pull(y),
     c("hello", "hello", "b", "hello", "c")
@@ -30,6 +37,13 @@ test_that("basic behavior of if_else() works", {
   expect_equal(
     test |>
       mutate(y = if_else(x1 == 'a', "foo", "bar")) |>
+      pull(y),
+    c("foo", "foo", "bar", "foo", "bar")
+  )
+
+  expect_equal(
+    test |>
+      mutate(y = dplyr::if_else(x1 == 'a', "foo", "bar")) |>
       pull(y),
     c("foo", "foo", "bar", "foo", "bar")
   )